got the python for the control module working (needs c code, and a decision on how to map this to c code)
