+++ spynnaker/pyNN/models/abstract_models/abstract_contains_a_synaptic_manager.py
@@ -1,0 +1,58 @@
+from abc import ABCMeta
+from six import add_metaclass
+
+from spynnaker.pyNN.models.neuron.synaptic_manager import SynapticManager
+
+
+@add_metaclass(ABCMeta)
+class AbstractContainsASynapticManager(object):
+
+    def __init__(self, synapse_type, ring_buffer_sigma, spikes_per_second):
+
+        # Set up synapse handling
+        self._synapse_manager = SynapticManager(
+            synapse_type, ring_buffer_sigma, spikes_per_second)
+
+    @property
+    def ring_buffer_sigma(self):
+        return self._synapse_manager.ring_buffer_sigma
+
+    @ring_buffer_sigma.setter
+    def ring_buffer_sigma(self, ring_buffer_sigma):
+        self._synapse_manager.ring_buffer_sigma = ring_buffer_sigma
+
+    @property
+    def spikes_per_second(self):
+        return self._synapse_manager.spikes_per_second
+
+    @spikes_per_second.setter
+    def spikes_per_second(self, spikes_per_second):
+        self._synapse_manager.spikes_per_second = spikes_per_second
+
+    @property
+    def synapse_dynamics(self):
+        return self._synapse_manager.synapse_dynamics
+
+    @synapse_dynamics.setter
+    def synapse_dynamics(self, synapse_dynamics):
+        self._synapse_manager.synapse_dynamics = synapse_dynamics
+
+    def add_pre_run_connection_holder(
+            self, connection_holder, edge, synapse_info):
+        self._synapse_manager.add_pre_run_connection_holder(
+            connection_holder, edge, synapse_info)
+
+    def get_connections_from_machine(
+            self, transceiver, placement, edge, graph_mapper,
+            routing_infos, synapse_info, machine_time_step):
+        return self._synapse_manager.get_connections_from_machine(
+            transceiver, placement, edge, graph_mapper,
+            routing_infos, synapse_info, machine_time_step)
+
+    @property
+    def synapse_type(self):
+        return self._synapse_manager.synapse_type
+
+    def get_maximum_delay_supported_in_ms(self, machine_time_step):
+        return self._synapse_manager.get_maximum_delay_supported_in_ms(
+            machine_time_step)
